Move operation fixtures to umami/core
